IPark-Gwanggyo/requirements.txt
flask
flask_wtf
flask_moment
flask_httpauth
flask_bootstrap
werkzeug
paho-mqtt
PyQt5
pyserial
requests
beautifulsoup4
